+++ src/main/java/insane96mcp/iguanatweaksreborn/module/combat/stats/Stats.java
@@ -103,7 +103,6 @@ public class Stats extends Feature {
 			MCUtils.applyModifier(player, Attributes.MOVEMENT_SPEED, MOVEMENT_SPEED_REDUCTION_UUID, "Movement Speed reduction", -playersMovementSpeedReduction, AttributeModifier.Operation.MULTIPLY_BASE, false);
 	}
 
-	//TODO Doesn't work properly with attributes on multiple slots
 	@SubscribeEvent(priority = EventPriority.HIGHEST)
 	public void onItemTooltipEvent(ItemTooltipEvent event) {
 		if (!this.isEnabled()