+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## 3.0.4
 * Added player's killer bounty
   * Mobs that kill players will not despawn and will drop 4x more experience and loot
 * Fixed diamond generation backport not working